+++ exasol_integration_test_docker_environment/lib/api/push_test_container.py
@@ -19,7 +19,7 @@ from exasol_integration_test_docker_environment.lib.config.build_config import (
 from exasol_integration_test_docker_environment.lib.config.docker_config import (
     set_docker_repository_config,
 )
-from exasol_integration_test_docker_environment.lib.data.test_container_content_description import (
+from exasol_integration_test_docker_environment.lib.models.data.test_container_content_description import (
     TestContainerContentDescription,
 )
 from exasol_integration_test_docker_environment.lib.docker.images.image_info import (
