+++ exasol_integration_test_docker_environment/lib/api/spawn_test_environment_with_test_container.py
@@ -30,10 +30,10 @@ from exasol_integration_test_docker_environment.lib.config.build_config import (
 from exasol_integration_test_docker_environment.lib.config.docker_config import (
     set_docker_repository_config,
 )
-from exasol_integration_test_docker_environment.lib.data.environment_info import (
+from exasol_integration_test_docker_environment.lib.models.data.environment_info import (
     EnvironmentInfo,
 )
-from exasol_integration_test_docker_environment.lib.data.test_container_content_description import (
+from exasol_integration_test_docker_environment.lib.models.data.test_container_content_description import (
     TestContainerContentDescription,
 )
 from exasol_integration_test_docker_environment.lib.docker.container.utils import (
