+++ exasol_integration_test_docker_environment/lib/models/data/database_credentials.py
