+++ exasol_integration_test_docker_environment/lib/models/data/docker_network_info.py
