+++ exasol_integration_test_docker_environment/lib/models/data/docker_volume_info.py
