+++ exasol_integration_test_docker_environment/lib/models/data/environment_info.py
@@ -4,16 +4,16 @@ from typing import (
 )
 
 from exasol_integration_test_docker_environment.lib.base.info import Info
-from exasol_integration_test_docker_environment.lib.data.container_info import (
+from exasol_integration_test_docker_environment.lib.models.data.container_info import (
     ContainerInfo,
 )
-from exasol_integration_test_docker_environment.lib.data.database_info import (
+from exasol_integration_test_docker_environment.lib.models.data.database_info import (
     DatabaseInfo,
 )
-from exasol_integration_test_docker_environment.lib.data.docker_network_info import (
+from exasol_integration_test_docker_environment.lib.models.data.docker_network_info import (
     DockerNetworkInfo,
 )
-from exasol_integration_test_docker_environment.lib.data.environment_type import (
+from exasol_integration_test_docker_environment.lib.models.data.environment_type import (
     EnvironmentType,
 )
