+++ exasol_integration_test_docker_environment/lib/models/data/environment_type.py
