+++ exasol_integration_test_docker_environment/lib/models/data/ssh_info.py
