+++ exasol_integration_test_docker_environment/lib/models/data/test_container_content_description.py
