+++ exasol_integration_test_docker_environment/lib/test_environment/database_waiters/wait_for_test_docker_database.py
@@ -18,10 +18,10 @@ from exasol_integration_test_docker_environment.lib.base.docker_base_task import
 from exasol_integration_test_docker_environment.lib.base.json_pickle_parameter import (
     JsonPickleParameter,
 )
-from exasol_integration_test_docker_environment.lib.data.database_credentials import (
+from exasol_integration_test_docker_environment.lib.models.data.database_credentials import (
     DatabaseCredentialsParameter,
 )
-from exasol_integration_test_docker_environment.lib.data.database_info import (
+from exasol_integration_test_docker_environment.lib.models.data import (
     DatabaseInfo,
 )
 from exasol_integration_test_docker_environment.lib.test_environment.database_waiters.db_container_log_thread import (
