+++ exasol_integration_test_docker_environment/lib/test_environment/parameter/test_container_parameter.py
@@ -3,7 +3,7 @@ from luigi.parameter import ParameterVisibility
 from exasol_integration_test_docker_environment.lib.base.json_pickle_parameter import (
     JsonPickleParameter,
 )
-from exasol_integration_test_docker_environment.lib.data.test_container_content_description import (
+from exasol_integration_test_docker_environment.lib.models.data.test_container_content_description import (
     TestContainerContentDescription,
 )
