+++ exasol_integration_test_docker_environment/lib/test_environment/spawn_test_environment.py
@@ -3,7 +3,7 @@ import luigi
 from exasol_integration_test_docker_environment.lib.base.dependency_logger_base_task import (
     DependencyLoggerBaseTask,
 )
-from exasol_integration_test_docker_environment.lib.data.environment_type import (
+from exasol_integration_test_docker_environment.lib.models.data.environment_type import (
     EnvironmentType,
 )
 from exasol_integration_test_docker_environment.lib.test_environment.parameter.spawn_test_environment_parameter import (
