+++ exasol_integration_test_docker_environment/lib/test_environment/spawn_test_environment_with_docker_db.py
@@ -6,19 +6,16 @@ from exasol_integration_test_docker_environment.lib.base.db_os_executor import (
     DockerExecFactory,
     SshExecFactory,
 )
-from exasol_integration_test_docker_environment.lib.data.container_info import (
-    ContainerInfo,
-)
-from exasol_integration_test_docker_environment.lib.data.database_info import (
+from exasol_integration_test_docker_environment.lib.models.data.database_info import (
     DatabaseInfo,
 )
-from exasol_integration_test_docker_environment.lib.data.docker_network_info import (
+from exasol_integration_test_docker_environment.lib.models.data.docker_network_info import (
     DockerNetworkInfo,
 )
-from exasol_integration_test_docker_environment.lib.data.docker_volume_info import (
+from exasol_integration_test_docker_environment.lib.models.data.docker_volume_info import (
     DockerVolumeInfo,
 )
-from exasol_integration_test_docker_environment.lib.data.environment_type import (
+from exasol_integration_test_docker_environment.lib.models.data.environment_type import (
     EnvironmentType,
 )
 from exasol_integration_test_docker_environment.lib.test_environment.abstract_spawn_test_environment import (
