+++ exasol_integration_test_docker_environment/test/get_test_container_content.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 from typing import Tuple
 
-from exasol_integration_test_docker_environment.lib.data.test_container_content_description import (
+from exasol_integration_test_docker_environment.lib.models.data.test_container_content_description import (
     TestContainerBuildMapping,
     TestContainerContentDescription,
     TestContainerRuntimeMapping,
