+++ exasol_integration_test_docker_environment/test/test_populate_data.py
@@ -10,7 +10,7 @@ import luigi
 from exasol_integration_test_docker_environment.lib.base.run_task import (
     generate_root_task,
 )
-from exasol_integration_test_docker_environment.lib.data.test_container_content_description import (
+from exasol_integration_test_docker_environment.lib.models.data.test_container_content_description import (
     TestContainerRuntimeMapping,
 )
 from exasol_integration_test_docker_environment.lib.docker import ContextDockerClient
