+++ exasol_integration_test_docker_environment/test/test_test_container_reuse.py
@@ -17,10 +17,10 @@ from exasol_integration_test_docker_environment.lib.config.build_config import (
 from exasol_integration_test_docker_environment.lib.config.docker_config import (
     set_docker_repository_config,
 )
-from exasol_integration_test_docker_environment.lib.data.container_info import (
+from exasol_integration_test_docker_environment.lib.models.data.container_info import (
     ContainerInfo,
 )
-from exasol_integration_test_docker_environment.lib.data.test_container_content_description import (
+from exasol_integration_test_docker_environment.lib.models.data.test_container_content_description import (
     TestContainerContentDescription,
 )
 from exasol_integration_test_docker_environment.lib.docker import ContextDockerClient
