+++ exasol_integration_test_docker_environment/testing/exaslct_docker_test_environment.py
@@ -4,7 +4,7 @@ from typing import Optional
 from exasol_integration_test_docker_environment.cli.options.test_environment_options import (
     LATEST_DB_VERSION,
 )
-from exasol_integration_test_docker_environment.lib.data.environment_info import (
+from exasol_integration_test_docker_environment.lib.models.data.environment_info import (
     EnvironmentInfo,
 )
 from exasol_integration_test_docker_environment.lib.test_environment.ports import Ports
